requirements.txt
ctransformers
pinecone-client
langchain
flask
pypdf
langchain_pinecone
langchain-community
langchain-huggingface
transformers
openai
tiktoken
python-dotenv

#installing setup.py
-e .  # this will look for setup.py file and run it